# feat(core): Lucene faceting for 300x more efficient label-value queries