--- a/core/src/main/resources/filodb-defaults.conf
+++ b/core/src/main/resources/filodb-defaults.conf
@@ -435,6 +435,9 @@ filodb {
     # At the cost of some extra heap memory, we can track queries holding shared lock for a long time
     # and starving the exclusive access of lock for eviction
     track-queries-holding-eviction-lock = true
+
+    # Whether or not in TimeSeriesShard, faceting is enabled for lucene index
+    index-faceting-enabled = true
   }
 
   # for standalone worker cluster configuration, see akka-bootstrapper