… core/src/test/resources/application_test.conf
… core/src/test/scala/filodb.core/memstore/PartKeyLuceneIndexSpec.scala