**Pull Request checklist**

- [x] The commit(s) message(s) follows the contribution [guidelines](CONTRIBUTING.md) ?
- [x] Tests for the changes have been added (for bug fixes / features) ?
- [ ] Docs have been added / updated (for bug fixes / features) ?

**Current behavior :**

For metadata searches that need to enumerate values of fields, we end up iterating through matching documents and build a hash set of field values. This is inefficient. It eats up CPU creating bottlenecks in query execution, especially in high cardinality clusters and reduces overall query throughput.

**New behavior :**

This PR uses Lucene faceting which tremendously improves performance (by 300x) of label-value queries which is a major CPU consumer in the system today.

* Faceting is enabled in raw cluster and disabled in downsample cluster
* Faceting can be disabled in raw cluster via configuration
* If faceting is enabled, label-value queries automatically choose the efficient execution path. Otherwise, they fall back to existing approach.
* Faceting requires 10% more disk space. So raw clusters will need more disk space.
* Faceting increases indexing time by 2x, so node bootstrap will take more time.


## Benchmark of Label-Value Query Performance

### With Synthetic Data 
Benchmarks performed with `PartKeyIndexBenchmark`
Num part keys: 1 mil
Part Keys generated using TestTimeSeriesProducer
All time series have same metric
Query to collect metric names has a ws/ns filter.

#### Prior to this PR, without faceting 
Index Size: 227MB
Time to index: 17s
All time series have about 25 unique metric names
Throughput of getting list of metric names (iterative): 0.281 ops/s

#### With this PR, with faceting all labels
Index Size: 242MB
Time to index: 35s
Throughput of getting list of metric names (facet collector): 113.276 ops/s

### With Production Data 
Data cannot be uploaded to git since it is not sanitized.
Benchmarks performed with `PartKeyIndexBenchmark`
Num part keys: 574k
Query to collect metric names has a ws/ns filter.

#### Prior to this PR, without faceting 
Index Size: 310MB
Time to index: 17s
Throughput of getting list of metric names (iterative): 0.065 ops/s


#### With this PR, with faceting all labels
Index Size: 337MB
Time to index: 39s
Throughput of getting list of metric names (facet collector): 26.4 ops/s

### Summary: 
Index size increased by 6% to 8%
Time to index increased by around 2x
Throughput of label values query increased by 300 to 400 times